# internal/api/docs/swagger.yaml
definitions:
...
  exercise.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  exercise.CreateExerciseRequest:
    properties:
      categoryID:
        type: integer
      description:
        type: string
      equipmentID:
        type: integer
      muscleGroupIDs:
        items:
          type: integer
        type: array
      name:
        type: string
      typeIDs:
        items:
          type: integer
        type: array
    type: object
  exercise.Equipment:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  exercise.Exercise:
    properties:
      category:
        $ref: '#/definitions/exercise.Category'
      createdAt:
        type: string
      description:
        type: string
      equipment:
        $ref: '#/definitions/exercise.Equipment'
      id:
        type: integer
      muscleGroups:
        items:
          $ref: '#/definitions/exercise.MuscleGroup'
        type: array
      name:
        type: string
      training_types:
        items:
          $ref: '#/definitions/exercise.TrainingType'
        type: array
      updatedAt:
        type: string
    type: object
  exercise.MuscleGroup:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  exercise.TrainingType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  exercise.UpdateExerciseRequest:
    properties:
      categoryID:
        type: integer
      description:
        type: string
      equipmentID:
        type: integer
      muscleGroupIDs:
        items:
          type: integer
        type: array
      name:
        type: string
      typeIDs:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/exercises:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/exercise.Exercise'
            type: array
      security:
      - BearerAuth: []
      summary: List all exercises
      tags:
      - exercises
    post:
      consumes:
      - application/json
      parameters:
      - description: Exercise payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/exercise.CreateExerciseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/exercise.Exercise'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new exercise
      tags:
      - exercises
  /api/v1/admin/exercises/{id}:
    delete:
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an exercise
      tags:
      - exercises
    get:
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/exercise.Exercise'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get exercise by ID
      tags:
      - exercises
    put:
      consumes:
      - application/json
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      - description: Exercise payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/exercise.UpdateExerciseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/exercise.UpdateExerciseRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an exercise
      tags:
      - exercises
  /api/v1/admin/exercises/{id}/details:
    get:
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/exercise.Exercise'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get exercise with full details
      tags:
      - exercises
  /api/v1/admin/exercises/by-muscle:
    get:
      parameters:
      - description: Muscle group name
        in: query
        name: muscle_group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/exercise.Exercise'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get exercises by muscle group
      tags:
      - exercises
  /api/v1/admin/exercises/by-training-type:
    get:
      parameters:
      - description: Training type name
        in: query
        name: training_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/exercise.Exercise'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get exercises by training type
      tags:
      - exercises
  /api/v1/admin/exercises/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/exercise.Exercise'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Search exercises
      tags:
      - exercises
  /api/v1/admin/muscle-groups:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/exercise.MuscleGroup'
            type: array
      security:
      - BearerAuth: []
      summary: List all muscle groups
      tags:
      - muscle-groups
...
securityDefinitions:
  BearerAuth:
    description: JWT-based authentication using Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"